+++ safekeeper/tests/walproposer_sim/block_storage.rs
@@ -1,52 +1,9 @@
 use std::collections::HashMap;
-use std::sync::Arc;
-
-use parking_lot::Mutex;
-use safekeeper::safekeeper::SafeKeeperState;
-use utils::id::TenantTimelineId;
-
-pub struct Disk {
-    pub timelines: Mutex<HashMap<TenantTimelineId, Arc<TimelineDisk>>>,
-}
-
-impl Default for Disk {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl Disk {
-    pub fn new() -> Self {
-        Disk {
-            timelines: Mutex::new(HashMap::new()),
-        }
-    }
-
-    pub fn put_state(&self, ttid: &TenantTimelineId, state: SafeKeeperState) -> Arc<TimelineDisk> {
-        self.timelines
-            .lock()
-            .entry(*ttid)
-            .and_modify(|e| {
-                let mut mu = e.state.lock();
-                *mu = state.clone();
-            })
-            .or_insert_with(|| {
-                Arc::new(TimelineDisk {
-                    state: Mutex::new(state),
-                    wal: Mutex::new(BlockStorage::new()),
-                })
-            })
-            .clone()
-    }
-}
-
-pub struct TimelineDisk {
-    pub state: Mutex<SafeKeeperState>,
-    pub wal: Mutex<BlockStorage>,
-}
 
 const BLOCK_SIZE: usize = 8192;
 
+/// A simple in-memory implementation of a block storage. Can be used to implement external
+/// storage in tests.
 pub struct BlockStorage {
     blocks: HashMap<u64, [u8; BLOCK_SIZE]>,
 }
